move scrollable tests to side panel
